+++ Assets/Scripts/Game/NoteData.cs
@@ -6,11 +6,11 @@ namespace SCOdyssey.Game
     public class NoteData
     {
         public int index;           // 채보 순서
-        public float time;          // 판정 시간
+        public double time;          // 판정 시간
         public NoteType noteType;   // 노트 타입
         public int laneIndex;     // 라인 번호
         
-        public NoteData(int index, float time, NoteType noteType, int laneIndex)
+        public NoteData(int index, double time, NoteType noteType, int laneIndex)
         {
             this.index = index;
             this.time = time;
